+++ style.css
@@ -68,7 +68,7 @@ body {
 
 .arenas {
     width: 786px;
-    height: 578px;
+    height: 70vh;
     background-size: cover;
     background-position: center;
     background-repeat: no-repeat;
@@ -420,9 +420,12 @@ input[type=radio]:checked + label > span {
     font-weight: 700;
     text-shadow: 0px 2px 2px rgba(0, 0, 0, 0.2);
 }
+input[type=radio]:disabled + label > span {
+    cursor: not-allowed;
+}
 input[type=radio]:checked + label > span::before {
     background: #f5d100;
-    box-shadow: 0px 3px 6px 0px rgba(0, 0, 0, 0.2);
+    box-shadow: 0 3px 6px 0 rgba(0, 0, 0, 0.2);
 }
 input[type=radio]:checked + label > span:hover {
     background: none;
